Increased map size and added a roof
No errors! Its really easy to change the map now. you need to follow
these rules when adding a square:
1) if you drag and drop from the project tab you must set the object to
a whole number before manipulating it (as in set it to 0,0) then when
you drag it around hold down cntrl and it will drag in increments of 1
square. (you can copy and paste existing squares and then the square
will already be on a whole number and you can skip that first part).
2) set the square to the right layer and tag. ***You must also set the
render to the right layer as well***. Also, squares that behave the same
but look different have a common layer (i.e. Walls and roofs of a
building both are in the wall layer).
3) if you increase the map size it must stay square. and you need to
manually change the mapSize variable in the CharacterMovement script.