Added Enemies
We have bugs!!!
added bugs to the map. I had to split up our old charactermovement scrpt
so now the mapping is taken care of by a separate object. The enemies
walk around and after each step by the player or by any of the enemies
the map is rebuilt.